refactor: move print_can_frame to usart.c
