+++ gripper_17.X/include/usart.h
@@ -425,12 +425,13 @@ bool SERCOM0_USART_ReceiverIsReady(void);
 int SERCOM0_USART_ReadByte(void);
 
 USART_ERROR SERCOM0_USART_ErrorGet(void);
-
-// static int USART_printf(char c, FILE *stream);
-//
-// static FILE USART_stream() = FDEV_SETUP_STREAM(USART_printf, NULL, _FDEV_SETUP_WRITE);
  
 uint32_t SERCOM0_USART_FrequencyGet(void);
+
+#ifdef DEBUG
+void print_can_frame(void);
+#endif
+
 // DOM-IGNORE-BEGIN
 #ifdef __cplusplus  // Provide C++ Compatibility
 }
